# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PREDICTA AI API
  version: 1.0.0
  summary: Canonical contract for the PREDICTA AI frontend and backend
  description: |
    Contract-first API for sports intelligence, calibrated model probabilities,
    deterministic value calculations and fact-grounded AI explanations.

    Probabilities are statistical estimates, never guarantees. `data_mode`
    distinguishes mock payloads from live/provider-backed payloads. Availability
    and freshness are carried by `DataQuality`; missing numeric data is `null`,
    never silently replaced by zero.
  contact:
    name: PREDICTA AI Architecture
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin deployment
security: []
...
paths:
  /dashboard:
    get:
      operationId: getDashboard
      tags: [Dashboard]
      summary: Get the aggregate dashboard snapshot
      description: |
        Returns the frontend dashboard payload in one read. Filtering by sport
        remains client-side because the dashboard includes the supported sports.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Dashboard snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /sports:
    get:
      operationId: getSports
      tags: [Catalog]
      summary: List supported sports
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Supported sports
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SportsEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /leagues:
    get:
      operationId: getLeagues
      tags: [Catalog]
      summary: List leagues
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered leagues
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /leagues/{league_id}:
    get:
      operationId: getLeague
      tags: [Catalog]
      summary: Get a league and its current available context
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDetailEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /teams:
    get:
      operationId: getTeams
      tags: [Catalog]
      summary: List teams or sport-side competitors
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered teams
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /teams/{team_id}:
    get:
      operationId: getTeam
      tags: [Catalog]
      summary: Get a team profile
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetailEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /players:
    get:
      operationId: getPlayers
      tags: [Catalog]
      summary: List players
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered players
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /players/{player_id}:
    get:
      operationId: getPlayer
      tags: [Catalog]
      summary: Get a player profile
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/PlayerId"
      responses:
        "200":
          description: Player detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetailEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /matches:
    get:
      operationId: getMatches
      tags: [Matches]
      summary: List matches
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/LeagueFilter"
        - $ref: "#/components/parameters/MatchDateFilter"
        - $ref: "#/components/parameters/MatchStatusFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered matches
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /matches/{match_id}:
    get:
      operationId: getMatch
      tags: [Matches]
      summary: Get the frontend match detail aggregate
      description: |
        Returns match context plus currently available statistics, odds,
        prediction and form. Missing sections are `null` or empty and are
        explained through `unavailable_fields`; they are never fabricated.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Match detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetailEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /matches/{match_id}/stats:
    get:
      operationId: getMatchStatistics
      tags: [Statistics]
      summary: Get normalized statistics for a match
      description: |
        Canonical statistics subresource. The current frontend obtains the same
        array embedded in `MatchDetail.stats`; both representations must be
        produced from the same backend DTO.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Match statistics
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchStatisticsEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /matches/{match_id}/odds:
    get:
      operationId: getMatchOdds
      tags: [Odds]
      summary: Get the latest compatible odds snapshot
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
        - $ref: "#/components/parameters/MarketFilter"
      responses:
        "200":
          description: Odds snapshot; data is null when unavailable
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OddsEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /matches/{match_id}/prediction:
    get:
      operationId: getMatchPrediction
      tags: [Predictions]
      summary: Get the published calibrated prediction
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
        - $ref: "#/components/parameters/MarketFilter"
      responses:
        "200":
          description: Prediction; data is null when no prediction is published
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /football/predictions/{match_id}:
    get:
      operationId: getFootballModelPrediction
      tags: [Predictions]
      summary: Get candidate football 1X2 model probabilities
      description: |
        Serves calibrated P(Home), P(Draw) and P(Away) from the versioned
        football Elo artefact currently registered for this route
        (`football-elo-v1-candidate`, `model_status=candidate`).

        This is a statistical estimate, never a certainty, and never a betting
        recommendation. Odds, expected value and the Value Engine are out of
        scope. Point-in-time Elo features are loaded from
        `football-1x2-history-0.3`; rows that are missing, mock, or only
        knowable after kickoff are refused.

        `GET /matches/{match_id}/prediction` is unchanged and remains the
        published frontend DTO. A future champion can replace the loaded
        artefact behind this path without changing the response schema.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
        - $ref: "#/components/parameters/CutoffAt"
      responses:
        "200":
          description: Calibrated 1X2 probabilities from the registered football model
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootballModelPredictionEnvelope"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/ValidationProblem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /football/value/{match_id}:
    get:
      operationId: getFootballMatchValue
      tags: [Value]
      summary: Evaluate football 1X2 value at a point-in-time cutoff
      description: |
        Combines the candidate Prediction Service probabilities with the latest
        complete odds snapshot satisfying `available_at <= cutoff_at`. The Value
        Engine never creates model probabilities. Odds, edge and EV are ratios,
        not recommendations or guarantees.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/MatchId"
        - $ref: "#/components/parameters/CutoffAt"
      responses:
        "200":
          description: Deterministic football 1X2 value analysis
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootballValueAnalysisEnvelope"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/ValidationProblem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /picks:
    get:
      operationId: getPicks
      tags: [Picks]
      summary: List published model picks
      description: Picks are model signals and never guaranteed outcomes.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/LeagueFilter"
        - $ref: "#/components/parameters/MatchDateFilter"
        - $ref: "#/components/parameters/MatchStatusFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered picks
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /value:
    get:
      operationId: getValueOpportunities
      tags: [Value]
      summary: List deterministic value evaluations
      description: |
        Every item is derived from a calibrated probability and a timestamped
        odds snapshot. Ratios are raw decimal values: `0.072` means 7.2%.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/SportFilter"
        - $ref: "#/components/parameters/LeagueFilter"
        - $ref: "#/components/parameters/MatchDateFilter"
        - $ref: "#/components/parameters/MatchStatusFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered value evaluations
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /performance:
    get:
      operationId: getPerformance
      tags: [Performance]
      summary: Get model performance and calibration
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Current model performance report
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
  /ai/analyze:
    post:
      operationId: analyzeMatch
      tags: [Analyst]
      summary: Get a fact-grounded analyst session
      description: |
        The analyst may explain only facts present in the validated fact pack.
        It cannot create probabilities, odds, statistics or availability data.
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalystRequest"
      responses:
        "200":
          description: Analyst session
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalystEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
    XRequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-provided correlation identifier. The API generates one when absent.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    SportFilter:
      name: sport
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/SportCode"
    LeagueFilter:
      name: league_id
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Identifier"
    MatchDateFilter:
      name: date
      in: query
      required: false
      description: Calendar date interpreted in the competition timezone.
      schema:
        type: string
        format: date
    MatchStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MatchStatus"
    SearchQuery:
      name: query
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 100
    MarketFilter:
      name: market
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 64
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items. The frontend may omit it.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      description: Zero-based offset. V1 list payloads retain `items` and `total`.
      schema:
        type: integer
        minimum: 0
        default: 0
    CutoffAt:
      name: cutoff_at
      in: query
      required: false
      description: |
        Exclusive-of-future PIT boundary. Omit to use kickoff (`pre_kickoff`).
        Values after kickoff are rejected as temporal leakage. Values before
        kickoff are rejected because this candidate only stores the pre-kickoff
        Elo snapshot.
      schema:
        $ref: "#/components/schemas/Timestamp"
    MatchId:
      name: match_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    LeagueId:
      name: league_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    TeamId:
      name: team_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    PlayerId:
      name: player_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
  headers:
    XRequestId:
      description: Correlation identifier, identical to the envelope or problem `request_id`.
      schema:
        type: string
  responses:
    Problem:
      description: Request or server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationProblem:
      description: Invalid request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
...
    SportCode:
      type: string
      enum: [football, basketball, tennis]
...
    MatchStatus:
      type: string
      enum: [scheduled, live, finished, postponed]
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, request_id]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: [string, "null"]
          format: uri-reference
        request_id:
          type: string
          minLength: 1
...
    EnvelopeMetadata:
      type: object
      required: [data_mode, generated_at, request_id]
      properties:
        data_mode:
          $ref: "#/components/schemas/DataMode"
        generated_at:
          $ref: "#/components/schemas/Timestamp"
        request_id:
          type: string
          minLength: 1
...
    Sport:
      type: object
      additionalProperties: false
      required: [id, name, code]
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        name:
          type: string
          minLength: 1
        code:
          $ref: "#/components/schemas/SportCode"
...
    DashboardSnapshot:
      type: object
      additionalProperties: false
      required:
        - headline
        - sports
        - matches_today
        - picks
        - value_opportunities
        - insights
        - model_health
      properties:
        headline:
          type: string
        sports:
          type: array
          items:
            $ref: "#/components/schemas/Sport"
        matches_today:
          type: array
          items:
            $ref: "#/components/schemas/MatchSummary"
        picks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
        value_opportunities:
          type: array
          items:
            $ref: "#/components/schemas/ValueOpportunity"
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
        model_health:
          $ref: "#/components/schemas/ModelHealthSummary"
...
    AnalystRequest:
      type: object
      additionalProperties: false
      required: [match_id, question]
      properties:
        match_id:
          $ref: "#/components/schemas/Identifier"
        question:
          type: [string, "null"]
          minLength: 1
          maxLength: 2000
    LeagueList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/League"
        total:
          type: integer
          minimum: 0
...
    DashboardEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DashboardSnapshot"
    SportsEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Sport"
    LeagueListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/LeagueList"
    LeagueDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/LeagueDetail"
    TeamListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/TeamList"
    TeamDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/TeamDetail"
    PlayerListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PlayerList"
    PlayerDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PlayerDetail"
    MatchListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/MatchList"
    MatchDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/MatchDetail"
    MatchStatisticsEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/MatchStatistic"
    OddsEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              anyOf:
                - $ref: "#/components/schemas/OddsSnapshot"
                - type: "null"
    PredictionEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              anyOf:
                - $ref: "#/components/schemas/PredictionDetail"
                - type: "null"
    FootballModelPredictionEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/FootballModelPrediction"
    FootballValueAnalysisEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/FootballValueAnalysis"
    PickListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PickList"
    ValueListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ValueList"
    PerformanceEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PerformanceReport"
    AnalystEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeMetadata"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AnalystSession"